Fix ambiguous scoping issue
In cases where the optional `to` argument is not provided to `slice()`, the scoping of `arguments.to` on line 881 fails because `arguments.to` is scoped too specifically. The unscoped `to` assignment on line 873 is thus not an overwriting of `arguments.to`, but instead the creation of a non-locally-scoped variable (in effect, `variables.to`, depending on the cfml engine being used and settings in effect). This change removes the ambiguity by locally var-scoping `calculatedTo` and `calculatedFrom`. The latter is technically unnecessary because the `from` argument includes a default value and therefore always exists, but I added it for symmetry and clarity.